This commit was generated by cvs2svn to compensate for changes in r13747, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13748 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/serial/serial-iec-lib.c
@@ -42,7 +42,7 @@ int serial_iec_lib_directory(unsigned int unit, const char *pattern,
 {
     int length, status;
     BYTE data, *tmpbuf;
-    unsigned int maxlen = 0;
+    size_t maxlen = 0;
 
     serial_iec_open(unit, 0, pattern, strlen(pattern));
